Represent FLAC's 64-bit fields as fixnum Int64 for web precision
Replace the byte-inspection placeholder workaround with a proper
Int64 sentinel so every 64-bit spec field parses losslessly on every
platform, not just the VM.

- Add package:fixnum ^1.1.0 as a runtime dependency.
- readUint64BE returns Int64 (via Int64.fromBytesBigEndian), eliminating
  the int-as-64-bit-unsigned trick that silently lost precision on
  JavaScript Number for values above 2^53.
- SeekPoint.sampleNumber / streamOffset, CueSheetTrack.trackOffset,
  CueSheetTrackIndex.offset, CueSheetBlock.leadInSamples change from
  int to Int64. The placeholder sentinel is now Int64(-1) (all-ones
  two's-complement), and SeekPoint.isPlaceholder compares against it
  directly.
- FlacReader.byteOffsetForSample calls .toInt() when combining Int64
  seek-point fields with the plain-int scanOffset/scanSample locals.
  Practical max is 2^53 bytes / samples — far beyond any real file.
- CueSheet test updated to unwrap Int64 values with .toInt() when
  asserting against plain int literals, and now also checks
  trackOffset on the lead-out track.
- README + CHANGELOG document the Int64 choice and why it's needed
  for web.

Verified again with `dart compile js`; full suite 87/87.

+++ lib/src/metadata/cue_sheet.dart
@@ -1,11 +1,13 @@
 import 'dart:typed_data';
 
+import 'package:fixnum/fixnum.dart';
+
 import 'metadata_block.dart';
 
 /// A single track in a [CueSheetBlock].
 class CueSheetTrack {
   /// Track offset in samples from the beginning of the audio stream.
-  final int trackOffset;
+  final Int64 trackOffset;
 
   /// Track number (1–99 for CD-DA, 1–254 for others; 170 for lead-out).
   final int trackNumber;
@@ -38,7 +40,7 @@ class CueSheetTrack {
 /// An index point within a [CueSheetTrack].
 class CueSheetTrackIndex {
   /// Sample offset of the index point relative to the track offset, in samples.
-  final int offset;
+  final Int64 offset;
 
   /// Index number (0 or 1 for the first two index points, etc.).
   final int indexNumber;
@@ -58,7 +60,7 @@ class CueSheetBlock extends MetadataBlock {
   final String mediaCatalogNumber;
 
   /// Number of lead-in samples (for CD-DA this is ≥ 2 seconds worth).
-  final int leadInSamples;
+  final Int64 leadInSamples;
 
   /// Whether the cue sheet corresponds to a CD.
   final bool isCD;
